+++ scripts/pack.ps1
@@ -1,4 +1,4 @@
-Remove-Item "package-release" -Recurse  -ErrorAction SilentlyContinue
+Remove-Item "package-release" -Recurse -ErrorAction SilentlyContinue
 New-Item "package-release" -ItemType Directory
 
 $exclude = @(
@@ -23,8 +23,7 @@ Copy-Item "LICENSE.md" -Destination "package-release/LICENSE.md"
 
 # Copy samples
 Copy-Item "samples/unity-of-bugs/Assets/Scenes" -Destination "package-release/Samples~/unity-of-bugs/Scenes" -Recurse
-Copy-Item "samples/unity-of-bugs/Assets/Scripts" -Destination "package-release/Samples~/unity-of-bugs/Scripts" -Recurse `
-    -Exclude "SmokeTestOptions.cs*"
+Copy-Item "samples/unity-of-bugs/Assets/Scripts" -Destination "package-release/Samples~/unity-of-bugs/Scripts" -Recurse
 
 # Create zip
 Compress-Archive "package-release/*" -DestinationPath "package-release.zip" -Force