+++ src/Sentry.Unity.Editor/ConfigurationWindow/OptionsConfigurationTab.cs
@@ -1,7 +1,10 @@
 using System;
 using System.IO;
+using System.Text;
 using UnityEditor;
 using UnityEngine;
+using RuntimeConfiguration = Sentry.Unity.ScriptableOptionsConfiguration;
+using BuildtimeConfiguration = Sentry.Unity.Editor.ScriptableOptionsConfiguration;
 
 namespace Sentry.Unity.Editor.ConfigurationWindow
 {
@@ -13,50 +16,72 @@ namespace Sentry.Unity.Editor.ConfigurationWindow
 
             EditorGUILayout.Space();
 
-            EditorGUILayout.HelpBox("The scriptable options configuration allows you to programmatically " +
-                                    "modify the Sentry options at runtime, right before the SDK gets initialized." +
-                                    "This allows you to override configuration otherwise unavailable from the " +
-                                    "Editor UI, e.g. set a custom BeforeSend callback.",
-                                    MessageType.Info);
+            EditorGUILayout.HelpBox(
+                "The scriptable options configuration allows you to programmatically modify Sentry options." +
+                "\n" +
+                "\n" +
+                "You can use 'Runtime Options Script' to modify options just before Sentry SDK gets initialized. " +
+                "This allows you to change configuration otherwise unavailable from the Editor UI, e.g. set a custom BeforeSend callback." +
+                "\n" +
+                "\n" +
+                "Use 'Buildtime Options Script' in case you need to change build-time behavior, e.g. specify custom Sentry-CLI options " +
+                "or change settings for native SDKs that start before the managed layer does (such as Android, iOS, macOS).",
+                MessageType.Info);
 
             EditorGUILayout.HelpBox("Clicking the 'New' button will prompt you with selecting a location for " +
-                                    "your custom 'ScriptableSentryConfiguration' script and automatically " +
-                                    "create a new instance at 'Assets/Resources/Sentry/'",
-                                    // TODO other platforms
-                                    // "Because Sentry Unity integration includes both managed C# Unity SDK and a " +
-                                    // "platform specific one, you can specify the respective overrides separately.\n\n" +
-                                    MessageType.Info);
+                                    "your custom 'ScriptableOptionsConfiguration' script and automatically " +
+                                    "create a new asset instance.", MessageType.Info);
 
             EditorGUILayout.Space();
 
-            OptionsConfigurationDotNet.Display(options);
+            options.OptionsConfiguration = OptionsConfigurationItem.Display(
+                options.OptionsConfiguration,
+                "Runtime Options Script",
+                "SentryRuntimeOptionsConfiguration"
+            );
+
+            options.BuildtimeOptionsConfiguration = OptionsConfigurationItem.Display(
+                options.BuildtimeOptionsConfiguration as BuildtimeConfiguration,
+                "Buildtime Options Script",
+                "SentryBuildtimeOptionsConfiguration"
+            ) as ScriptableObject;
         }
     }
 
-    internal static class OptionsConfigurationDotNet
+    internal static class OptionsConfigurationItem
     {
         private const string CreateScriptableObjectFlag = "CreateScriptableOptionsObject";
         private const string ScriptNameKey = "ScriptableOptionsName";
 
-        public static void Display(ScriptableSentryUnityOptions options)
+        public static T? Display<T>(T? value, string fieldName, string scriptName) where T : ScriptableObject
         {
             GUILayout.BeginHorizontal();
-            options.OptionsConfiguration = EditorGUILayout.ObjectField(
-                    new GUIContent(".NET (C#)", "A scriptable object that inherits from " +
-                                                            "'ScriptableOptionsConfiguration' and allows you to " +
-                                                            "programmatically modify Sentry options."),
-                    options.OptionsConfiguration, typeof(ScriptableOptionsConfiguration), false)
-                as ScriptableOptionsConfiguration;
+            var result = EditorGUILayout.ObjectField(
+                    new GUIContent(fieldName, "A scriptable object that inherits from 'ScriptableOptionsConfiguration' " +
+                                         "and allows you to programmatically modify Sentry options."),
+                    value,
+                    typeof(T),
+                    false
+                ) as T;
             if (GUILayout.Button("New", GUILayout.ExpandWidth(false)))
             {
-                CreateScript();
+                CreateScript<T>(fieldName, scriptName);
             }
             GUILayout.EndHorizontal();
+            return result;
         }
 
-        internal static void CreateScript()
+        private static void CreateScript<T>(string fieldName, string scriptName)
         {
-            var scriptPath = EditorUtility.SaveFilePanel("Sentry Options Configuration", "Assets", "SentryOptionsConfiguration", "cs");
+            var isEditorScript = typeof(T) == typeof(BuildtimeConfiguration);
+            var directory = isEditorScript ? "Assets/Editor" : "Assets/Scripts";
+
+            if (!AssetDatabase.IsValidFolder(directory))
+            {
+                AssetDatabase.CreateFolder(Path.GetDirectoryName(directory), Path.GetFileName(directory));
+            }
+
+            var scriptPath = EditorUtility.SaveFilePanel(fieldName, directory, scriptName, "cs");
             if (String.IsNullOrEmpty(scriptPath))
             {
                 return;
@@ -68,37 +93,30 @@ namespace Sentry.Unity.Editor.ConfigurationWindow
                 scriptPath = "Assets" + scriptPath.Substring(Application.dataPath.Length);
             }
 
-            var scriptName = Path.GetFileNameWithoutExtension(scriptPath);
+            scriptName = Path.GetFileNameWithoutExtension(scriptPath);
 
-            File.WriteAllText(scriptPath, $@"using System;
-using Sentry.Unity;
-using UnityEngine;
+            var template = new StringBuilder();
+            template.AppendLine("using System;");
+            template.AppendLine("using UnityEngine;");
+            template.AppendLine("using Sentry.Unity;");
+            if (isEditorScript)
+            {
+                template.AppendLine("using Sentry.Unity.Editor;");
+            }
+            template.AppendLine();
+            template.AppendFormat("[CreateAssetMenu(fileName = \"Assets/Resources/Sentry/{0}.cs\", menuName = \"Sentry/{0}\", order = 999)]\n", scriptName);
+            template.AppendFormat("public class {0} : {1}\n", scriptName, typeof(T).FullName);
+            template.AppendLine("{");
+            template.AppendLine("    /// See base class for documentation.");
+            template.AppendLine("    /// Learn more at https://docs.sentry.io/platforms/unity/configuration/options/#programmatic-configuration");
+            template.AppendFormat("    public override void Configure(SentryUnityOptions options{0})\n",
+                                  typeof(T) == typeof(BuildtimeConfiguration) ? ", SentryCliOptions cliOptions" : "");
+            template.AppendLine("    {");
+            template.AppendLine("        // TODO implement");
+            template.AppendLine("    }");
+            template.AppendLine("}");
 
-/// Learn more at https://docs.sentry.io/platforms/unity/configuration/options/#programmatic-configuration
-[CreateAssetMenu(fileName = ""Assets/Resources/Sentry/{scriptName}.cs"", menuName = ""Sentry/{scriptName}"", order = 999)]
-public class {scriptName} : ScriptableOptionsConfiguration
-{{
-    /// See base class for documentation.
-    public override void ConfigureAtBuild(SentryUnityOptions options, SentryCliOptions? cliOptions)
-    {{
-        Configure(options);
-        // TODO implement buildtime-specific configuration, if any
-    }}
-
-    /// See base class for documentation.
-    public override void ConfigureAtRuntime(SentryUnityOptions options)
-    {{
-        Configure(options);
-        // TODO implement runtime-specific configuration, if any
-    }}
-
-    /// Usually, it's a good idea to have common configuration and only call really specific setup in the above methods.
-    private void Configure(SentryUnityOptions options)
-    {{
-        // TODO implement
-    }}
-}}
-");
+            File.WriteAllText(scriptPath, template.ToString().Replace("\r\n", "\n"));
 
             // The created script has to be compiled and the scriptable object can't immediately be instantiated.
             // So instead we work around this by setting a 'CreateScriptableObjectFlag' flag in the EditorPrefs to
@@ -131,11 +149,24 @@ public class {scriptName} : ScriptableOptionsConfiguration
             AssetDatabase.CreateAsset(optionsConfigurationObject, $"Assets/Resources/Sentry/{scriptName}.asset");
             AssetDatabase.Refresh();
 
-            // Don't overwrite already set OptionsConfiguration
             var options = EditorWindow.GetWindow<SentryWindow>().Options;
-            if (options.OptionsConfiguration is null)
+
+            if (optionsConfigurationObject is RuntimeConfiguration)
             {
-                options.OptionsConfiguration = (ScriptableOptionsConfiguration)optionsConfigurationObject;
+                // Don't overwrite already set OptionsConfiguration
+                if (options.OptionsConfiguration is null)
+                {
+                    options.OptionsConfiguration = optionsConfigurationObject as RuntimeConfiguration;
+                }
+            }
+
+            if (optionsConfigurationObject is BuildtimeConfiguration)
+            {
+                // Don't overwrite already set OptionsConfiguration
+                if (options.BuildtimeOptionsConfiguration is null)
+                {
+                    options.BuildtimeOptionsConfiguration = optionsConfigurationObject;
+                }
             }
         }
     }